+++ back/src/companies/sirene/searchCompanies.ts
@@ -1,10 +1,9 @@
 import { searchCompanies as searchCompaniesInsee } from "./insee/client";
 import { searchCompanies as searchCompaniesDataGouv } from "./entreprise.data.gouv.fr/client";
 import { searchCompanies as searchCompaniesSocialGouv } from "./social.gouv/client";
+import { searchCompanies as searchCompaniesTD } from "./trackdechets/client";
 import { backoffIfTooManyRequests, throttle } from "./ratelimit";
 import { redundant } from "./redundancy";
-
-const { INSEE_MAINTENANCE } = process.env;
 
 const searchCompaniesInseeThrottled = backoffIfTooManyRequests(
   searchCompaniesInsee,
@@ -26,7 +25,8 @@ const searchCompaniesSocialGouvThrottled = throttle(searchCompaniesSocialGouv, {
 // list different implementations of searchCompanies by
 // order of priority.
 const searchCompaniesProviders = [
-  ...(INSEE_MAINTENANCE === "true" ? [] : [searchCompaniesInseeThrottled]),
+  searchCompaniesTD,
+  searchCompaniesInseeThrottled,
   searchCompaniesDataGouvThrottled,
   searchCompaniesSocialGouvThrottled
 ];