+++ search/src/commands/indexInseeSiret.ts
@@ -12,7 +12,7 @@ import {
 } from "../indexation/elasticSearch.helpers";
 
 process.on("exit", function () {
-  console.log(`Command indexInseeSirene.ts finished`);
+  console.log("Command indexInseeSirene.ts finished");
   logger.end();
 });
 
@@ -34,6 +34,9 @@ const siretWithUniteLegaleFormatter = async (
   body: ElasticBulkNonFlatPayload,
   extras: { sireneIndexConfig: IndexProcessConfig }
 ): Promise<ElasticBulkNonFlatPayload> => {
+  if (!body.length) {
+    return [];
+  }
   const response = await multiGet(body, extras.sireneIndexConfig);
   return response.body.docs.map((sirenDoc, i) => [
     body[i][0],
@@ -52,7 +55,7 @@ const siretUrl =
   "https://files.data.gouv.fr/insee-sirene/StockEtablissement_utf8.zip";
 
 const siretIndexConfig: IndexProcessConfig = {
-  alias: `stocketablissement_utf8${INDEX_ALIAS_NAME_SEPARATOR}${
+  alias: `stocketablissement${INDEX_ALIAS_NAME_SEPARATOR}${
     process.env.NODE_ENV ? process.env.NODE_ENV : "dev"
   }${
     process.env.INDEX_ALIAS_NAME_SUFFIX
@@ -63,10 +66,75 @@ const siretIndexConfig: IndexProcessConfig = {
   csvFileName: "StockEtablissement_utf8.csv",
   // zip target filename
   zipFileName: "StockEtablissement_utf8.zip",
-  idKey: "siren",
-  // append StockUniteLegale by JOINING ON "siren"
+  idKey: "siret",
+  // append StockUniteLegale by JOINING ON siren
   dataFormatterFn: siretWithUniteLegaleFormatter,
-  mappings: standardMapping
+  // copy_to full-text search field to optimize multiple field search performance
+  // docs https://www.elastic.co/guide/en/elasticsearch/reference/7.16/copy-to.html
+  mappings: {
+    _doc: {
+      // inherit from standardMapping
+      ...standardMapping._doc,
+      // override
+      properties: {
+        siren: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        siret: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        denominationUniteLegale: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        nomUniteLegale: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        denominationUsuelleEtablissement: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        denominationUsuelle1UniteLegale: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        denominationUsuelle2UniteLegale: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        denominationUsuelle3UniteLegale: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        nomUsageUniteLegale: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        sigleUniteLegale: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        enseigne1Etablissement: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        enseigne2Etablissement: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        enseigne3Etablissement: {
+          type: "text",
+          copy_to: "td_search_companies"
+        },
+        td_search_companies: {
+          type: "text"
+        }
+      }
+    }
+  }
 };
 
 /**
@@ -79,5 +147,5 @@ const siretIndexConfig: IndexProcessConfig = {
   } else {
     await downloadAndIndex(siretUrl, siretIndexConfig);
   }
-  logger.info(`Command indexInseeSiret.ts finished`);
+  logger.info("Command indexInseeSiret.ts finished");
 })();